requirements.txt
discord.py==1.7.2
lyricsgenius==3.0.0